Move helper function into memqueue
